Move encode decode multisig functions to common
